+++ android/app/src/main/kotlin/com/example/simple_dyphic/MainActivity.kt
@@ -1,6 +1,5 @@
-package jp.hotdrop.esdyphic
+package com.example.simple_dyphic
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity: FlutterActivity()
