+++ pubspec.yaml
@@ -6,7 +6,7 @@ publish_to: 'none'
 version: 3.0.0+4
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -33,7 +33,7 @@ dependencies:
 
   # other
   package_info_plus: ^4.1.0
-  intl: ^0.18.0
+  intl: ^0.19.0
   logger: ^1.1.0
 
 dev_dependencies:
